+++ src/main/java/fr/openmc/core/features/dream/registries/items/armors/creaking/OldCreakingBoots.java
@@ -12,7 +12,7 @@ public class OldCreakingBoots extends DreamEquipableItem {
     }
 
     @Override
-    public int getAdditionalMaxTime() {
+    public long getAdditionalMaxTime() {
         return 5;
     }
