+++ src/main/java/fr/openmc/core/features/dream/registries/items/armors/soul/SoulChestplate.java
@@ -12,7 +12,7 @@ public class SoulChestplate extends DreamEquipableItem {
     }
 
     @Override
-    public int getAdditionalMaxTime() {
+    public long getAdditionalMaxTime() {
         return 15;
     }
